added mobile top-bar functionality
