+++ client/src/assets/scss/main.scss
@@ -6,6 +6,8 @@
 @import "./stylesheets/ProfilePage";
 @import "./stylesheets/Word-Tile";
 
+@import "./stylesheets/top-bar";
+
 @include foundation-everything;
 // place global css here
 *,
@@ -14,38 +16,26 @@
   box-sizing: border-box;
 }
 
-.logo {
-  max-width: 5em;
-  align-items: center;
-  height: auto;
+.word-smith {
+  font-size: x-large;
+  color: orange;
 }
-.logo:hover {
+
+.link {
+  font-size: x-large;
+  color:black;
+  // background-color: #EDE0D4;
+  // border: solid 1px rgb(209, 163, 111);
+  // border-radius: 5px;
+  // margin-left: 1em;
+}
+
+.link:hover {
   opacity: 80%;
+  color:rgb(255, 0, 0);
+  // background-color: #e9d6c3;
 }
 
-.sign-in{
-  margin-left: 0;
-  border-radius: .2em;
-
-}
-.top-bar {
-  padding: 1em;
-
-  .top-bar-left{
-    max-width: 30%;
-  }
-  .top-bar-right{
-    margin: auto 0;
-
-    .profilebtn {
-      margin-top: .75em;
-    }
-  }
-  .sign-out{
-    margin-top: .75em;
-    padding: .45em;
-  }
-}
 
 body{
   background-color: #eccfb2;
