+++ client/src/assets/scss/stylesheets/HomePage.scss
@@ -7,37 +7,30 @@
   margin: 2em;
 }
 
-
 .home-buttons{
   align-items: center;
   justify-content: center;
   display: flex;
+  padding: 2em;
 
-  .filter-btn {
+  .button-style {
     font-size: large;
-    margin-left: 3em;
     padding: .75em;
     border-radius: 1em;
     background-color: #EDE0D4;
     border: solid 1px rgb(209, 163, 111);
     border-radius: 5px;
   }
-  .add-btn {
-    margin-top: auto;
-    font-size: large;
-    margin-right: 3em;
-    padding: .75em;
-    border-radius: 1em;
-    background-color: #EDE0D4;
-    border: solid 1px rgb(209, 163, 111);
-    border-radius: 5px;
+  .button-style:hover {
+    color:rgb(255, 0, 0);
+    background-color: #e9d6c3;
+  }
+
+  .new-filter-form {
+    display: flex;
   }
 }
 
-.add-btn:hover, .filter-btn:hover {
-  color:rgb(255, 0, 0);
-  background-color: #e9d6c3;
-}
 
 .filter-label{
   font-size: x-large;
@@ -75,33 +68,7 @@
   text-align: center;
 }
 
-// .filter-form {
-//   text-align: center;
-//   border-radius: 1em;
-//   background-color: #EDE0D4;
-//   border: solid 1px rgb(209, 163, 111);
-//   border-radius: 5px;
-//   margin-left: 30%;
-//   margin-right: 30%;
-//   margin-top: 1em;
-//   padding-bottom: 1em;
-//   padding-top: 1em;
-// }
-
-.filterForm{
-  margin-left: 10em;
-}
-
 .select-new{
   width: fit-content;
-}
-.new-filter-form {
-  margin-top: 2em;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-  display: flex;
-}
-.filter-submit{
-  margin-left: 2em;
+  max-width: 18em;
 }
