+++ client/src/assets/scss/stylesheets/LandingPage.scss
@@ -1,15 +1,13 @@
-.link {
-  font-size: x-large;
-  color:black;
-  background-color: #EDE0D4;
-  border: solid 1px rgb(209, 163, 111);
-  border-radius: 5px;
-  margin-left: 1em;
-}
-
-.link:hover {
-  color:rgb(255, 0, 0);
-  background-color: #e9d6c3;
+.landing-main {
+  width: 100vw;
+  align-items: center;
+  text-align: center;
+  padding-top: 100px;
+  color: white;
+  overflow: hidden;
+  background-color: #eccfb2;
+  background-size: cover;
+  background-repeat: no-repeat;
 }
 
 .main-page-redirect {
@@ -33,15 +31,3 @@
   margin-bottom: 2em;
 }
 
-.main {
-  height:1500px;
-  width: 100vw;
-  align-items: center;
-  text-align: center;
-  padding-top: 100px;
-  color: white;
-  overflow: hidden;
-  background-color: #eccfb2;
-  background-size: cover;
-  background-repeat: no-repeat;
-}
